requirements.txt
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
rich==13.9.4
openai==1.61.1
setuptools==75.8.0
argparse==1.4.0
fastapi==0.115.8
fastapi-cli==0.0.7
firecrawl-py==1.12.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-text-splitters==0.3.6